Migrate to ES6 classes
