+++ types/battle.js
