+++ types/core.js
