+++ types/entity.js
